scroll delta was unsigned short

+++ private/win/window.c
@@ -283,7 +283,7 @@ static LRESULT window_proc(HWND const hwnd, UINT const msg, WPARAM const wparam,
                     if(mouse->usButtonFlags & RI_MOUSE_WHEEL) {
                         platform->key_state[MIMAS_MOUSE_WHEEL] = mouse->usButtonData;
                         if(_mimas->active_window && _mimas->active_window->callbacks.scroll) {
-                            _mimas->active_window->callbacks.scroll(_mimas->active_window, 0, mouse->usButtonData, _mimas->active_window->callbacks.scroll_data);
+                            _mimas->active_window->callbacks.scroll(_mimas->active_window, 0, (short)mouse->usButtonData, _mimas->active_window->callbacks.scroll_data);
                         }
                         // Reset because there's no event for 0 delta.
                         platform->key_state[MIMAS_MOUSE_WHEEL] = 0;
@@ -292,7 +292,7 @@ static LRESULT window_proc(HWND const hwnd, UINT const msg, WPARAM const wparam,
                     if(mouse->usButtonFlags & RI_MOUSE_HWHEEL) {
                         platform->key_state[MIMAS_MOUSE_HORIZ_WHEEL] = mouse->usButtonData;
                         if(_mimas->active_window && _mimas->active_window->callbacks.scroll) {
-                            _mimas->active_window->callbacks.scroll(_mimas->active_window, mouse->usButtonData, 0, _mimas->active_window->callbacks.scroll_data);
+                            _mimas->active_window->callbacks.scroll(_mimas->active_window, (short)mouse->usButtonData, 0, _mimas->active_window->callbacks.scroll_data);
                         }
                         // Reset because there's no event for 0 delta.
                         platform->key_state[MIMAS_MOUSE_HORIZ_WHEEL] = 0;
